+++ Magic/src/main/java/com/elmakers/mine/bukkit/entity/EntityData.java
@@ -735,6 +735,10 @@ public class EntityData implements com.elmakers.mine.bukkit.api.entity.EntityDat
             if (apiMage instanceof com.elmakers.mine.bukkit.magic.Mage) {
                 ((com.elmakers.mine.bukkit.magic.Mage)apiMage).setEntityData(this);
             }
+
+            if (mageData != null) {
+                mageData.onSpawn(apiMage);
+            }
         }
 
         return true;